+++ config/environments/test.rb
@@ -16,11 +16,9 @@ Rails.application.configure do
   # system, or in some way before deploying your code.
   config.eager_load = ENV['CI'].present?
 
-  # Configure public file server for tests with Cache-Control for performance.
-  config.public_file_server.enabled = true
-  config.public_file_server.headers = {
-    'Cache-Control' => "public, max-age=#{1.hour.to_i}"
-  }
+  # Disable serving static files from the `/public` folder by default since
+  # we don't have any, except public.html which is served by AuthController#index
+  config.public_file_server.enabled = false
 
   # Show full error reports and disable caching.
   config.consider_all_requests_local       = true