+++ CHANGES.md
@@ -3,6 +3,7 @@
 - Add `/closures` endpoint
 - Prevent reservations during closures
 - Include closure info in item metadata
+- Allow serving a test UI page when `SERVE_TEST_UI` is set
 
 # 0.0.1 (2022-09-23)